+++ revdep/checks/ordr/new/ordr.Rcheck/tests/testthat.R
@@ -1,4 +1,0 @@
-library(testthat)
-library(ordr)
-
-test_check("ordr")