+++ NEWS.md
@@ -1,8 +1,17 @@
+## Version 1.7.1 (2024-05-03)
+
+* Extended the "datasets" vignette with further description of how this was done.
+* Added `dogfood` and `oral` datasets, both simple MANOVA examples.
+
 ## Version 1.7.0 (2024-05-01)
+
+This is a semi-major release, primarily documenting all datasets in the package with `@concept` tags.
+It also adds new datasets and fixes bugs and documentation errors since the last CRAN release (v 1.6.2)
 
 * Added a new example (`AddHealth` data) to the `HE_manova` vignette
 * Added `Overdose` data, a simple MMRA example.
 * All datasets now classified by `@concept` tags in documentation, giving the primary methods ("MANOVA", "MMRA", "repeated", ...) that they illustrate. These supplement `@keyword` tags and appear in the index of the package documentation.
+* A new vignette (`datasets`) lists the datasets in the package, classified by these `@concept` tags.
 
 ## Version 1.6.3 (2024-03-05)